Create error boundary between header and archive index
